refactor: update import paths for History and Setting to chromeStorage directory
